Support gcc and armclang with CMake build

+++ SampleCode/ISP/ISP_SPI/targetdev.c
@@ -12,7 +12,7 @@
 // Supports 64K/128K/256K (APROM)
 uint32_t GetApromSize()
 {
-    uint32_t size = 0x10000, data;
+    unsigned int size = 0x10000, data;
     int result;
 
     do
@@ -21,7 +21,7 @@ uint32_t GetApromSize()
 
         if(result < 0)
         {
-            return size;
+            return (uint32_t)size;
         }
         else
         {
@@ -34,7 +34,7 @@ uint32_t GetApromSize()
 #define CONFIG0_DFEN                0x01
 void GetDataFlashInfo(uint32_t *addr, uint32_t *size)
 {
-    uint32_t uData;
+    unsigned int uData;
     *size = 0;
     FMC_Read_User(Config0, &uData);
